+++ samples/sprues/_advanced_sprues.rb
@@ -1,4 +1,5 @@
-require 'squib'
+# require 'squib'
+require_relative '../../lib/squib'
 
 # This is a more advanced example of Sprues
 # (mostly so we can have a torturous regression test)
@@ -14,10 +15,10 @@ Squib::Deck.new(cards: 6, width: '2.2in', height: '2.1in') do
 
   circle x: width/2, y: height/2, radius: 20 # midpoint
 
-  text str: (0..9).map{ |i| "Card #{i}\n\n" },
+  text str: (0..9).map{ |i| "Card! #{i}\n\n" },
        font: 'Sans 32', align: :center, valign: :middle,
        height: :deck, width: :deck
-  
+
   save_sheet sprue: 'my_sprues/weird_sprue.yml',
              trim: '0.2in', trim_radius: '10pt',
              prefix: "advanced_sprues_"